refactor: centralize log level api type
