+++ webui/react/src/pages/TaskLogs.settings.ts
@@ -1,6 +1,5 @@
 import { BaseType, SettingsConfig } from 'hooks/useSettings';
-
-import { LogLevelFromApi } from './TaskLogFilters';
+import { LogLevelFromApi } from 'types';
 
 export interface Settings {
   agentId?: string[];
