+++ webui/react/src/types.ts
@@ -614,6 +614,16 @@ export enum LogLevel {
   Warning = 'warning',
 }
 
+export enum LogLevelFromApi {
+  Unspecified = 'LOG_LEVEL_UNSPECIFIED',
+  Trace = 'LOG_LEVEL_TRACE',
+  Debug = 'LOG_LEVEL_DEBUG',
+  Info = 'LOG_LEVEL_INFO',
+  Warning = 'LOG_LEVEL_WARNING',
+  Error = 'LOG_LEVEL_ERROR',
+  Critical = 'LOG_LEVEL_CRITICAL',
+}
+
 export interface Log {
   id: number | string;
   level?: LogLevel;
